chore(docs): finish doctests and docstrings for `heimdall_common`
