+++ common/src/ether/evm/core/types.rs
@@ -318,7 +318,7 @@ pub fn convert_bitmask(instruction: Instruction) -> (usize, Vec<String>) {
 /// 2. Potential types that the byte size could be.
 ///
 /// ```
-/// use heimdall_common::ether::evm::utils::byte_size_to_type;
+/// use heimdall_common::ether::evm::core::types::byte_size_to_type;
 ///
 /// let (byte_size, potential_types) = byte_size_to_type(1);
 /// assert_eq!(byte_size, 1);
