+++ common/src/utils/http.rs
@@ -9,17 +9,20 @@ static APP_USER_AGENT: &str = concat!(env!("CARGO_PKG_NAME"), "/", env!("CARGO_P
 
 /// Make a GET request to the target URL and return the response body as JSON
 ///
-/// # Arguments
-/// `url` - the URL to make the GET request to
-/// `timeout` - timeout duration in seconds
+/// ```no_run
+/// use heimdall_common::utils::http::get_json_from_url;
 ///
-/// # Returns
-/// `Result<Option<Value>, reqwest::Error>` - the response body as JSON
+/// let url = "https://example.com";
+/// let timeout = 5;
+/// // get_json_from_url(url, timeout).await;
+/// ```
 pub async fn get_json_from_url(url: &str, timeout: u64) -> Result<Option<Value>, reqwest::Error> {
     _get_json_from_url(url, 0, 5, timeout).await
 }
 
 #[async_recursion]
+/// Internal function for making a GET request to the target URL and returning the response body as
+/// JSON
 async fn _get_json_from_url(
     url: &str,
     retry_count: u8,
