+++ common/src/utils/testing/mod.rs
@@ -1,2 +1,1 @@
 pub mod benchmarks;
-mod tests;
