+++ common/src/utils/version.rs
@@ -9,6 +9,7 @@ pub struct Version {
     pub patch: u32,
 }
 
+/// get the current version from cargo
 pub fn current_version() -> Version {
     // get the current version from the cargo package
     let version_string = env!("CARGO_PKG_VERSION");
@@ -21,6 +22,7 @@ pub fn current_version() -> Version {
     }
 }
 
+/// get the latest version from github
 pub async fn remote_version() -> Version {
     // get the latest release from github
     let remote_repository_url =
@@ -54,41 +56,132 @@ impl Display for Version {
 }
 
 impl Version {
-    // greater than
+    /// greater than
     pub fn gt(&self, other: &Version) -> bool {
         self.major > other.major ||
             (self.major == other.major && self.minor > other.minor) ||
             (self.major == other.major && self.minor == other.minor && self.patch > other.patch)
     }
 
-    // greater than or equal to
+    /// greater than or equal to
     pub fn gte(&self, other: &Version) -> bool {
         self.major > other.major ||
             (self.major == other.major && self.minor > other.minor) ||
             (self.major == other.major && self.minor == other.minor && self.patch >= other.patch)
     }
 
-    // less than
+    /// less than
     pub fn lt(&self, other: &Version) -> bool {
         self.major < other.major ||
             (self.major == other.major && self.minor < other.minor) ||
             (self.major == other.major && self.minor == other.minor && self.patch < other.patch)
     }
 
-    // less than or equal to
+    /// less than or equal to
     pub fn lte(&self, other: &Version) -> bool {
         self.major < other.major ||
             (self.major == other.major && self.minor < other.minor) ||
             (self.major == other.major && self.minor == other.minor && self.patch <= other.patch)
     }
 
-    // equal to
+    /// equal to
     pub fn eq(&self, other: &Version) -> bool {
         self.major == other.major && self.minor == other.minor && self.patch == other.patch
     }
 
-    // not equal to
+    /// not equal to
     pub fn ne(&self, other: &Version) -> bool {
         self.major != other.major || self.minor != other.minor || self.patch != other.patch
     }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::utils::version::*;
+
+    #[test]
+    fn test_greater_than() {
+        let v1 = Version { major: 2, minor: 3, patch: 4 };
+        let v2 = Version { major: 2, minor: 3, patch: 3 };
+        let v3 = Version { major: 2, minor: 2, patch: 5 };
+        let v4 = Version { major: 1, minor: 4, patch: 4 };
+
+        assert!(v1.gt(&v2));
+        assert!(v1.gt(&v3));
+        assert!(v1.gt(&v4));
+        assert!(!v2.gt(&v1));
+        assert!(!v1.gt(&v1));
+    }
+
+    #[test]
+    fn test_greater_than_or_equal_to() {
+        let v1 = Version { major: 2, minor: 3, patch: 4 };
+        let v2 = Version { major: 2, minor: 3, patch: 4 };
+
+        assert!(v1.gte(&v2));
+        assert!(v2.gte(&v1));
+        assert!(v1.gte(&Version { major: 1, minor: 0, patch: 0 }));
+    }
+
+    #[test]
+    fn test_less_than() {
+        let v1 = Version { major: 2, minor: 3, patch: 4 };
+        let v2 = Version { major: 2, minor: 3, patch: 5 };
+        let v3 = Version { major: 2, minor: 4, patch: 4 };
+        let v4 = Version { major: 3, minor: 3, patch: 4 };
+
+        assert!(v1.lt(&v2));
+        assert!(v1.lt(&v3));
+        assert!(v1.lt(&v4));
+        assert!(!v2.lt(&v1));
+        assert!(!v1.lt(&v1));
+    }
+
+    #[test]
+    fn test_less_than_or_equal_to() {
+        let v1 = Version { major: 2, minor: 3, patch: 4 };
+        let v2 = Version { major: 2, minor: 3, patch: 4 };
+
+        assert!(v1.lte(&v2));
+        assert!(v2.lte(&v1));
+        assert!(v1.lte(&Version { major: 3, minor: 0, patch: 0 }));
+    }
+
+    #[test]
+    fn test_equal_to() {
+        let v1 = Version { major: 2, minor: 3, patch: 4 };
+        let v2 = Version { major: 2, minor: 3, patch: 4 };
+        let v3 = Version { major: 2, minor: 3, patch: 5 };
+
+        assert!(v1.eq(&v2));
+        assert!(!v1.eq(&v3));
+    }
+
+    #[test]
+    fn test_not_equal_to() {
+        let v1 = Version { major: 2, minor: 3, patch: 4 };
+        let v2 = Version { major: 2, minor: 3, patch: 5 };
+        let v3 = Version { major: 3, minor: 3, patch: 4 };
+
+        assert!(v1.ne(&v2));
+        assert!(v1.ne(&v3));
+        assert!(!v1.ne(&Version { major: 2, minor: 3, patch: 4 }));
+    }
+
+    #[test]
+    fn test_version_display() {
+        let version = Version { major: 2, minor: 3, patch: 4 };
+
+        assert_eq!(version.to_string(), "2.3.4");
+    }
+
+    #[test]
+    fn test_version_current() {
+        let version = current_version();
+
+        assert_eq!(version.to_string(), env!("CARGO_PKG_VERSION"));
+    }
+
+    #[tokio::test]
+    async fn test_version_remote() {}
 }
